requirements.txt
Babel==2.9.1
kmeans_pytorch==0.3
matplotlib==3.3.4
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
rsa==4.7.2
torch==1.8.1
torchvision==0.9.1
